Getting ready to do mixed Swift & C code by linking to libpng.dylib and using a bridging header to make C code visible to Swift.
This is all due to the fact that Quartz Image IO does not support writing 1 bpp PNG files. So we do it the hard way.

The header file zlib.h is in the SDK.
The header files png.h, pngconf.h, and pnglibconf.h were fetched using Homebrew.
The libpng.dylib file was also fetched with Homebrew. Version 1.6.17 is being used.

These files were copied over to the project so that they could be found by the bundle. In this case, it shouldn't really be necessary. But it does keep the entire project together.